+++ na-frontend/js/components/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {connect} from 'react-redux';
 import Main from './Main';
 import NowPlaying from './NowPlaying';
-import {updateQueue, updatePlaying} from '../actions/actions';
+import {updatePlaying} from '../actions/actions';
 
 const divStyles = {
     display: 'flex',
@@ -14,14 +14,10 @@ class App extends React.Component {
     render() {
         return (
             <div style={divStyles}>
-                <Main queue={this.props.queue} refreshQueue={this.refreshQueue.bind(this)}/>
+                <Main/>
                 <NowPlaying playing={this.props.playing} refreshPlaying={this.refreshPlaying.bind(this)}/>
             </div>
         );
-    }
-
-    refreshQueue() {
-        this.props.dispatch(updateQueue());
     }
 
     refreshPlaying() {
@@ -31,8 +27,6 @@ class App extends React.Component {
 
 function select(state) {
     return {
-        loadingQueue: state.queue.loadingQueue,
-        queue: state.queue.queue,
         loadingPlaying: state.playing.loadingPlaying,
         playing: state.playing.playing
     };
