+++ na-frontend/js/components/Sidebar.js
@@ -2,13 +2,13 @@ import React from 'react';
 
 const divStyles = {
     border: '1px solid grey',
-    flex: 8
+    flex: 2
 };
 
-export default class Library extends React.Component {
+export default class Sidebar extends React.Component {
     render() {
         return (
-            <div style={divStyles}> Right</div>
+            <div style={divStyles}></div>
         );
     }
 }
